Fix bug 4385 - arithmetic overflow in cost estimates

+++ latest9.9/hej/net/sf/saxon/expr/Assignation.java
@@ -288,9 +288,10 @@ public abstract class Assignation extends Expression implements LocalBinding {
      * and we assume that a sequence has length 5. The resulting estimates may be used, for
      * example, to reorder the predicates in a filter expression so cheaper predicates are
      * evaluated first.
+     * @return
      */
     @Override
-    public int getCost() {
+    public double getCost() {
         return getSequence().getCost() + 5 * getAction().getCost();
     }
