+++ latest9.9/hej/net/sf/saxon/expr/LetExpression.java
@@ -410,9 +410,10 @@ public class LetExpression extends Assignation implements TailCallReturner {
      * and we assume that a sequence has length 5. The resulting estimates may be used, for
      * example, to reorder the predicates in a filter expression so cheaper predicates are
      * evaluated first.
+     * @return
      */
     @Override
-    public int getCost() {
+    public double getCost() {
         return getSequence().getCost() + getAction().getCost();
     }
