+++ latest9.9/hej/net/sf/saxon/expr/LookupExpression.java
@@ -184,9 +184,10 @@ public class LookupExpression extends BinaryExpression {
      * and we assume that a sequence has length 5. The resulting estimates may be used, for
      * example, to reorder the predicates in a filter expression so cheaper predicates are
      * evaluated first.
+     * @return
      */
     @Override
-    public int getCost() {
+    public double getCost() {
         return getLhsExpression().getCost() * getRhsExpression().getCost();
     }
